fix(score-progression): resolve code-review findings (round 1)
- Fix SVG gradient ID collision: DeltaChart now uses React.useId() so
  each instance gets a unique gradient ID; fixes wrong team colors when
  multiple delta charts are open simultaneously
- Fix as-cast: replace `e.target.value as ChartType` with an explicit
  type guard (value === "progression" || value === "delta")
- Split ProgressionChart and DeltaChart into their own sibling folders
  (progression-chart/, delta-chart/) per one-component-per-file rule
- Simplify buildScoreDelta: track prevScore as a variable to eliminate
  the points.at(-1) array scan and remove the unused teamLines param;
  precompute String(teamId0/1) outside the event loop
- Remove inline derivation comments from formatter tests
