+++ pages/src/components/stats/score-progression/score-progression-formatter.ts
@@ -15,31 +15,29 @@ type KillRaceEvent = NonNullable<MatchAnalytics["scoreProgression"]>["timeline"]
 function buildScoreDelta(
   teamIds: readonly number[],
   events: readonly KillRaceEvent[],
-  teamLines: readonly ScoreProgressionTeamLine[],
   durationMs: number,
 ): ScoreDeltaData | null {
-  if (teamIds.length < 2 || teamLines.length < 2) {
+  if (teamIds.length < 2) {
     return null;
   }
 
   const [teamId0, teamId1] = teamIds;
-  const [line0, line1] = teamLines;
+  const key0 = String(teamId0);
+  const key1 = String(teamId1);
 
   const points: ScoreProgressionPoint[] = [{ timestampMs: 0, score: 0 }];
+  let prevScore = 0;
 
   for (const event of events) {
-    const score0 = event.runningScores[String(teamId0)] ?? 0;
-    const score1 = event.runningScores[String(teamId1)] ?? 0;
-    const prevScore = points.at(-1)?.score ?? 0;
+    const score0 = event.runningScores[key0] ?? 0;
+    const score1 = event.runningScores[key1] ?? 0;
     const newScore = score0 - score1;
     points.push({ timestampMs: event.timestampMs, score: prevScore });
     points.push({ timestampMs: event.timestampMs, score: newScore });
+    prevScore = newScore;
   }
 
-  points.push({
-    timestampMs: durationMs,
-    score: (line0.points.at(-1)?.score ?? 0) - (line1.points.at(-1)?.score ?? 0),
-  });
+  points.push({ timestampMs: durationMs, score: prevScore });
 
   const scores = points.map((p) => p.score);
   const minScore = Math.min(...scores);
@@ -98,5 +96,5 @@ export function formatScoreProgression(
     teamLines.push({ teamId, name: state.name, color: state.color, points: state.points });
   }
 
-  return { durationMs, teamLines, scoreDelta: buildScoreDelta(teamIds, events, teamLines, durationMs) };
+  return { durationMs, teamLines, scoreDelta: buildScoreDelta(teamIds, events, durationMs) };
 }
